Allow python app to read PG connection from envvar
Modify the python app and docker-0compose settings so that the Postgres
connection can be read from environment variables, rather than
hardcoded. This most importantly allows the database containers for the
testing environments to be named differently, so that the development
container can still be up at the same time as the test containers.